Further split off the rest of the json stuff (- WIP #316 -)
Namely JsonLicenseFinderTable into its own module (dependant on JsonPrettyTable)
