+++ pip-licenses/output/__init__.py
@@ -42,33 +42,7 @@ from prettytable import (
 
 
 from .JsonPrettyTable import JsonPrettyTable
-
-
-class JsonLicenseFinderTable(JsonPrettyTable):
-    def format_row(self, row: RowType) -> dict[str, str | list[str]]:
-        resrow: dict[str, str | list[str]] = {}
-        for field, value in zip(self._field_names, row):
-            if field == "Name":
-                resrow["name"] = value
-
-            if field == "Version":
-                resrow["version"] = value
-
-            if field == "License":
-                resrow["licenses"] = [value]
-
-        return resrow
-
-    def get_string(self, **kwargs: str | list[str]) -> str:
-        # import included here in order to limit dependencies
-        # if not interested in JSON output,
-        # then the dependency is not required
-        import json
-
-        options = self._get_options(kwargs)
-        rows = self._get_rows(options)
-        lines = [self.format_row(row) for row in rows]
-        return json.dumps(lines, sort_keys=True)
+from .JsonLicenseFinderTable import JsonLicenseFinderTable
 
 
 class CSVPrettyTable(PrettyTable):
